refactor(sdk-metrics): use same benchmarks library for all scripts
