+++ packages/sdk-metrics/test/performance/benchmark/micro-benchmarks.js
@@ -1,0 +1,6 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+require('../benchmarks/otel-sdk-metrics-paths');
